requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
streamlit>=1.28.0
boto3>=1.34.0
hypothesis>=6.88.0
pytest>=7.4.0
python-dotenv>=1.0.0